+++ app/search/[keyword]/SearchResult.tsx
@@ -9,7 +9,7 @@ import { useQuery } from "@tanstack/react-query";
 import { docsQuery } from "@/services/docs/docs.query";
 import { useRouter } from "next/navigation";
 import { theme } from "@/styles";
-import { useDate } from "@/hooks/useDate";
+import { useDate } from "@/hooks";
 import { MoonLoader } from "react-spinners";
 import * as styles from "./style.css";
