+++ hooks/useDate.ts
@@ -2,7 +2,7 @@ import dayjs from "dayjs";
 import moment from "moment";
 import "moment/locale/ko";
 
-export const useDate = () => {
+const useDate = () => {
   const fromNow = (date: Date) => {
     return moment(date).fromNow();
   };
@@ -22,3 +22,5 @@ export const useDate = () => {
 
   return { fromNow, formatDate, getValidYearList };
 };
+
+export default useDate;
